pyproject.toml
[project]
name = "kwisser-readme"
version = "0.1.0"
description = "GitHub profile README generator for kwisser"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "lxml",
    "pydantic-settings>=2.13.1",
    "python-dateutil",
    "python-dotenv",
    "requests",
]

[project.scripts]
kwisser = "kwisser.__main__:main"

[tool.uv]
package = true
